feat: get nfts by users address

+++ interface/ICErc20.sol
@@ -1,5 +1,5 @@
+//SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
-
 
 interface ICErc20 {
     function mint(uint256) external returns (uint256);
@@ -8,8 +8,7 @@ interface ICErc20 {
 
     function supplyRatePerBlock() external returns (uint256);
 
-    function redeem(uint) external returns (uint);
+    function redeem(uint256) external returns (uint256);
 
-    function redeemUnderlying(uint) external returns (uint);
-
+    function redeemUnderlying(uint256) external returns (uint256);
 }
